+++ CHANGELOG.md
@@ -12,6 +12,8 @@ FlatLaf Change Log
 - If a JButton has an icon and no text, then it does not get a minimum width
   (usually 72 pixel) and the left and right insets are same as top/bottom insets
   so that it becomes square (if the icon is square).
+- Changed styling of default button in "Flat Light" theme (wide blue border
+  instead of blue background).
 - Added Java 9 module descriptor `module-info.class` to `flatlaf.jar` (in
   `META-INF/versions/9`). But FlatLaf remains Java 8 compatible. (issue #1)
 - Support specifying custom scale factor in system properties `flatlaf.uiScale`